+++ src/main.rs
@@ -7,15 +7,46 @@ use cli::{Ferium, ProfileSubCommands, SubCommands};
 use colored::{ColoredString, Colorize};
 use ferinth::Ferinth;
 use furse::Furse;
+use itertools::Itertools;
 use lazy_static::lazy_static;
 use libium::{add, config, upgrade};
 use tokio::{
-	fs::{create_dir_all, remove_dir_all},
+	fs::{create_dir_all, remove_file},
 	io::AsyncReadExt,
 };
 
+const CROSS: &str = "×";
 lazy_static! {
 	pub static ref TICK: ColoredString = "✓".green();
+}
+
+struct Downloadable {
+	filename: String,
+	download_url: String,
+}
+impl From<furse::structures::file_structs::File> for Downloadable {
+	fn from(file: furse::structures::file_structs::File) -> Self {
+		Self {
+			filename: file.file_name,
+			download_url: file.download_url,
+		}
+	}
+}
+impl From<ferinth::structures::version_structs::VersionFile> for Downloadable {
+	fn from(file: ferinth::structures::version_structs::VersionFile) -> Self {
+		Self {
+			filename: file.filename,
+			download_url: file.url,
+		}
+	}
+}
+impl From<octocrab::models::repos::Asset> for Downloadable {
+	fn from(asset: octocrab::models::repos::Asset) -> Self {
+		Self {
+			filename: asset.name,
+			download_url: asset.url.into(),
+		}
+	}
 }
 
 #[tokio::main]
@@ -38,7 +69,7 @@ async fn actual_main() -> Result<()> {
 		eprint!("Checking internet connection... ");
 		match online::check(Some(4)).await {
 			Ok(_) => println!("{}", *TICK),
-			Err(_) => bail!("× Ferium requires an internet connection to work"),
+			Err(_) => bail!("{} Ferium requires an internet connection to work", CROSS),
 		}
 	};
 
@@ -49,8 +80,7 @@ async fn actual_main() -> Result<()> {
 		"A CurseForge API key is required to build. If you don't have one, you can bypass this by setting the variable to a blank string, however anything using the CurseForge API will not work."
 	));
 	let mut config_file =
-		config::get_config_file(cli_app.config_file.unwrap_or_else(config::config_file_path))
-			.await?;
+		config::get_file(cli_app.config_file.unwrap_or_else(config::file_path)).await?;
 	let mut config_file_contents = String::new();
 	config_file
 		.read_to_string(&mut config_file_contents)
@@ -81,7 +111,7 @@ async fn actual_main() -> Result<()> {
 		.await?;
 
 		// Update config file and quit
-		config::write_config(&mut config_file, &config).await?;
+		config::write_file(&mut config_file, &config).await?;
 		return Ok(());
 	}
 
@@ -94,7 +124,7 @@ async fn actual_main() -> Result<()> {
 		}
 		// Default to first profile if index is set incorrectly
 		config.active_profile = 0;
-		config::write_config(&mut config_file, &config).await?;
+		config::write_file(&mut config_file, &config).await?;
 		bail!("Active profile specified incorrectly. Switched to first profile",)
 	};
 
@@ -169,61 +199,92 @@ async fn actual_main() -> Result<()> {
 		SubCommands::Sort => profile.mods.sort_by_cached_key(|mod_| mod_.name.clone()),
 		SubCommands::Upgrade { no_patch_check } => {
 			check_empty_profile(profile)?;
-			// Empty the mods directory
-			let _ = remove_dir_all(&profile.output_dir).await;
 			create_dir_all(&profile.output_dir).await?;
+			let mut to_download = Vec::new();
 			let mut error = false;
+
+			println!("{}\n", "Determining the Latest Compatible Versions".bold());
 			for mod_ in &profile.mods {
 				use libium::config::structs::ModIdentifier;
-				let result = match &mod_.identifier {
+				let result: Result<Downloadable, _> = match &mod_.identifier {
 					ModIdentifier::CurseForgeProject(project_id) => {
-						match upgrade::curseforge(&curseforge, profile, *project_id, no_patch_check)
+						upgrade::curseforge(&curseforge, profile, *project_id, no_patch_check)
 							.await
-						{
-							Ok(file) => Ok(file.file_name),
-							Err(err) => Err(err),
-						}
+							.map(std::convert::Into::into)
 					},
 					ModIdentifier::ModrinthProject(project_id) => {
-						match upgrade::modrinth(&modrinth, profile, project_id, no_patch_check)
+						upgrade::modrinth(&modrinth, profile, project_id, no_patch_check)
 							.await
-						{
-							Ok(version) => Ok(version.files[0].filename.clone()),
-							Err(err) => Err(err),
-						}
+							.map(|ok| ok.files[0].clone().into())
 					},
 					ModIdentifier::GitHubRepository(full_name) => {
-						match upgrade::github(&github.repos(&full_name.0, &full_name.1), profile)
+						upgrade::github(&github.repos(&full_name.0, &full_name.1), profile)
 							.await
-						{
-							Ok(asset) => Ok(asset.name),
-							Err(err) => Err(err),
-						}
+							.map(std::convert::Into::into)
 					},
 				};
 				match result {
-					Ok(file_name) => {
+					Ok(result) => {
 						println!(
 							"{} {:40}{}",
 							*TICK,
 							mod_.name,
-							format!("({})", file_name).dimmed()
+							format!("({})", result.filename).dimmed()
 						);
+						to_download.push(result);
 					},
 					Err(err) => {
-						eprintln!("{}", format!("× {:40}{}", mod_.name, err).red());
+						eprintln!("{}", format!("{} {:40}{}", CROSS, mod_.name, err).red());
 						error = true;
 					},
 				}
 			}
+
+			eprint!("\n{}", "Downloading Mod Files... ".bold());
+			for file in std::fs::read_dir(&profile.output_dir)? {
+				let file = file?;
+				let path = file.path();
+				if path.is_file() {
+					let mut index = None;
+					// If a file is already downloaded
+					if let Some(downloadable) = to_download
+						.iter()
+						.find_position(|thing| file.file_name().to_str().unwrap() == thing.filename)
+					{
+						index = Some(downloadable.0);
+					}
+					match index {
+						// Then don't download the file
+						Some(index) => {
+							to_download.swap_remove(index);
+						},
+						// Or else delete the file
+						None => remove_file(path).await?,
+					}
+				}
+			}
+			match {
+				for downloadable in to_download {
+					let contents = reqwest::get(downloadable.download_url)
+						.await?
+						.bytes()
+						.await?;
+					upgrade::write_mod_file(profile, contents, &downloadable.filename).await?;
+				}
+				Ok::<(), anyhow::Error>(())
+			} {
+				Ok(_) => println!("{}", *TICK),
+				Err(_) => bail!("{}", CROSS),
+			}
+
 			if error {
-				bail!("\nSome mods were not successfully downloaded")
+				bail!("\nCould not get the latest compatible version of some mods")
 			}
 		},
 	};
 
 	// Update config file with new values
-	config::write_config(&mut config_file, &config).await?;
+	config::write_file(&mut config_file, &config).await?;
 
 	Ok(())
 }